Add migration page to concepts menu
Without a link from the menu, it's very hard to discover this page. I only became aware of it from a [forum post](https://forum.holochain.org/t/updating-the-source-code-is-going-to-create-a-new-chain/7885). As a newbie to hApps, the absence of a page discussing how to upgrade code was very concerning in light of all the other statements I found about code hash changes creating a new chain:

> One of the most important characteristics of a DNA is that it's identified by the hash of the source-code. It follows that if you change something from that code and install that DNA again, you are literally creating a new network.